requirements.txt

numpy>=1.26
opencv-python>=4.9
pydantic>=2.8
PyYAML>=6.0
scipy>=1.12
pylepton>=1.0.0
